Task 012: Port FileSegment/Metadata + finish priority .cpp (sub-attempt S2)
Add FileSegment.cpp, Metadata.cpp, and FileCacheErrnoException.h, and finish the
five Task-011 priority/eviction .cpp against the real center-SCC headers. All
seven target TUs pass -fsyntax-only (no link/test — the green gate is S4;
FileCache.cpp/QueryLimit.cpp are S3).

Apply the four Controller/user-authorized structural resolutions:
- B1: EvictionCandidates.h portability (C++23 deducing-this -> const/non-const
  begin/end overloads; folly::F14FastSet::merge -> insert-range loop).
- B2a: inject reserve_space_wait_lock_timeout_milliseconds from FileCacheConfig
  into CacheMetadata (per design 04/08), replacing the global Context read.
- B2b: opened-file-handle invalidation is throw-not-implemented (VELOX_NYI) with
  a TODO(Task 013) at the rename and removal sites, raised OUTSIDE the
  best-effort try/catch so it propagates loudly; OpenedFileCache is not ported.
- B3: inject the manager/FileCache-owned FileCacheWorkerPool into CacheMetadata
  by reference (not owned), per thread-pool design 04.

FileSegment::write reconciles a partial physical append only on ENOSPC/EDQUOT
and rethrows the original exception; cancel-before-join, the ordered std::map
KeyMetadata, the F14 no-escape invariant, and LockedKey destruction order are
preserved.

+++ velox/ch/Interpreters/FileCache/EvictionCandidates.cpp
@@ -285,7 +285,7 @@ void EvictionCandidates::evict()
             chassert(
                 false, fmt::format(
                     "Failed to lock key {} (state: {}), but had {} eviction candidates from it",
-                    key, key_candidates.key_metadata->getState(), key_candidates.candidates.size()));
+                    key, static_cast<int>(key_candidates.key_metadata->getState()), key_candidates.candidates.size()));
             continue;
         }
